move checkbox component to ui library
