+++ packages/ui/src/components/CheckboxRx/index.ts
@@ -1,0 +1,4 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export * from './CheckboxRx';
