+++ packages/ui/src/components/CheckboxRx/types.ts
@@ -1,0 +1,10 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import type { AnyData } from '@polkadot-live/types/misc';
+
+export interface CheckboxRxProps {
+  selected: boolean;
+  theme: AnyData;
+  onChecked: () => void;
+}
