+++ packages/ui/src/components/index.ts
@@ -5,6 +5,7 @@ export * from './Accordion';
 export * from './AccordionRx';
 export * from './ActionItem';
 export * from './Cards';
+export * from './CheckboxRx';
 export * from './CopyButton';
 export * from './Common';
 export * from './DragClose';
